apps/parser-worker-python/requirements.txt
redis==5.0.1
spacy==3.7.2
requests==2.31.0
psycopg2-binary==2.9.9
